esg/services/embeddings/pyproject.toml
[project]
name = "embeddings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.59",
    "google-genai>=1.46.0",
    "langchain-google-genai>=3.0.0",
    "langchain-text-splitters>=1.0.0",
    "pdfplumber>=0.11.7",
    "pika>=1.3.2",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.26.5",
    "tqdm>=4.67.1",
]
